\name{Curated pathways and metadata}
\alias{Curated pathways and metadata}

\title{curated biological pathways}

\description{
    An \code{".xlsx"} file containing curated informations of 13 biological pathways and metadata: cell cycle,
    Hippo, Myc, Notch, Nrf2, Pi3k, TGF-Beta, Rtk-Ras, Tp53, Wnt, MutSig genes, OncoKB-CNAs-AMP, OncoKB-CNAs-HOMODEL. 
    Every first column of each electronic sheet contains HUGO symbols of every gene in the patwhay. 
    In addition, every pathway is enriched with several metadata informations
    (e.g., Alteration, Oncogenicity, PMIDs for mutation effect, aliases, GISTIC del, etc).
}