\name{"epi" GSE5847 RData Object}
\alias{"epi" GSE5847 RData Object}
\docType{data}
\title{
Epithelium transcriptome profile.
}
\description{
LCM dataset with matched stroma transcriptome profiles from the same tumor (GSE5847).
}
\usage{load(epi_GSE5847.RData)}
\format{
  A data frame with 3101 genes x 34 samples.
}
\details{
The dataset is automatically loaded with crossWGCNA library loading.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
Savino, A., de Marzo, N., Provero, P. & Poli, V. Meta-analysis of microdissected breast tumors reveals genes regulated in the stroma but hidden in bulk analysis. Cancers (Basel) 13, (2021).
}
\examples{
load(epi_GSE5847.RData)
}
\keyword{datasets}
