\name{"stroma" GSE5847 RData Object}
\alias{"stroma" GSE5847 RData Object}
\docType{data}
\title{
Stroma transcriptome profile.
}
\description{
LCM dataset with matched epithelium transcriptome profiles from the same tumor (GSE5847).
}
\usage{load(stroma_GSE5847.RData)}
\format{
  A data frame with 3101 genes x 34 samples.
}
\details{
The dataset is automatically loaded with crossWGCNA library loading.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
Savino, A., de Marzo, N., Provero, P. & Poli, V. Meta-analysis of microdissected breast tumors reveals genes regulated in the stroma but hidden in bulk analysis. Cancers (Basel) 13, (2021).
}
\examples{
load(stroma_GSE5847.RData)
## maybe str("stroma" GSE5847 RData Object) ; plot("stroma" GSE5847 RData Object) ...
}
\keyword{datasets}
