\name{"stroma" GSE5847 RData Object}
\alias{"stroma" GSE5847 RData Object}
\docType{data}
\title{
Stroma transcriptome profile.
}
\description{
LCM dataset with matched epithelium transcriptome profiles from the same tumor (GSE5847).
}
\usage{load("stroma_GSE5847.RData")}
\format{
  A data frame with \code{3101} genes \code{x} \code{34} samples.
}
\details{
The dataset is automatically loaded with crossWGCNA library loading.
}
\source{
code{\url{https://www.mdpi.com/2072-6694/13/13/3371}}
}
\references{
Savino, A., de Marzo, N., Provero, P. & Poli, V. Meta-analysis of microdissected breast tumors reveals genes regulated in the stroma but hidden in bulk analysis. Cancers (Basel) 13, (2021).
}
\examples{
load("stroma_GSE5847.RData")
}
\keyword{datasets}
